// src/adapters/api/login/dtos/login_password.dto.ts
// --------------------------------
// Helpers
// --------------------------------

class LoginPasswordDto {
  photoUrl: string;
  passwordPiece: string;
  photoRelativePath: string;
  photoName: string;
}

// --------------------------------
// Public Interface
// --------------------------------

export { LoginPasswordDto };